fix windows compile errors
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@8169 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_sim.cpp
@@ -1304,7 +1304,7 @@ void sim_t::analyze_player( player_t* p )
       }
       p -> dps_convergence_error[i] /= i;
       p -> dps_convergence_error[i] = sqrt( p -> dps_convergence_error[i] );
-      p -> dps_convergence_error[i] = 2.0 * p -> dps_convergence_error[i] / sqrt ( i );
+      p -> dps_convergence_error[i] = 2.0 * p -> dps_convergence_error[i] / sqrt ( (float) i );
 
     }
   }
